fix for money values in loans

+++ src/HE.Investments.Common/Messages/ValidationErrorMessage.cs
@@ -106,9 +106,9 @@ public static class ValidationErrorMessage
 
     public static string MustProvideYourLowerNumber(string fieldName, int maxValue) => $"Your {fieldName} must be {maxValue} or fewer";
 
-    public static string MustProvideTheHigherNumber(string fieldName, int minValue) => $"The {fieldName} must be {minValue} or more";
+    public static string MustProvideTheHigherNumber(string fieldName, long minValue) => $"The {fieldName} must be {minValue} or more";
 
-    public static string MustProvideTheLowerNumber(string fieldName, int maxValue) => $"The {fieldName} must be {maxValue} or fewer";
+    public static string MustProvideTheLowerNumber(string fieldName, long maxValue) => $"The {fieldName} must be {maxValue} or fewer";
 
     public static string MustProvideYourWholeNumber(string fieldName, int example = 30) => $"Your {fieldName} must be a whole number, like {example}";
